Refactor: Organize test files and enhance CMake configuration
- Moved test executables to a dedicated `test` directory for better structure.
- Updated CMakeLists.txt to reflect new paths for test source files.
- Added a comprehensive full VCF test report documenting execution and results.
- Introduced a new test for phases 1 and 2 functionality, ensuring core components are validated.
- Enhanced existing tests for phases 3, 4, and 5 to improve coverage and reliability.

+++ include/utils/FastaReader.hpp
@@ -56,7 +56,7 @@ public:
      * @param chr Chromosome name.
      * @return Length in bp, or -1 if chromosome not found.
      */
-    int64_t get_chr_length(const std::string& chr) const;
+    int32_t get_chr_length(const std::string& chr) const;
     
     /**
      * @brief Checks if the FASTA file was successfully loaded.
